pyproject.toml
[tool.poetry]
name = "puppet"
version = "0.1.0"
description = "Puppet is a discord bot developped in python which is a card collectible game in multiple universe."
authors = ["Kamigaku <kamigaku@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.14.8"
"discord.py" = {path = "../discord_py_japan7"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
